Split into files for brevity & add lots of comments
